pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "econ-clean"
version = "0.3.0"
description = "Comprehensive toolkit for economics panel data analysis"
readme = "docs/README.md"
authors = [
    {name = "Anton Ebsen"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "statsmodels>=0.13.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "python-calamine>=0.6.1",
    "pandera>=0.14.0",
    "pyyaml>=6.0",
    "streamlit>=1.20.0",
    "jinja2>=3.1.0",
    "openpyxl>=3.1.0",
    "pyarrow>=10.0.0",
    "fastparquet>=2023.0",
    "linearmodels>=4.25",
    "stargazer>=0.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "econ-clean[dev]",
]

[project.urls]
Homepage = "https://github.com/AntonEbsen/economics-of-the-welfare-state"
Documentation = "https://github.com/AntonEbsen/economics-of-the-welfare-state/tree/main/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
